Call OpenAPIClient from Service

+++ goyotashi/BusinessLogic/Entities/GroupSummary.swift
@@ -6,7 +6,7 @@
 //
 
 struct GroupSummary {
-    let groupId: Int
+    let groupId: Int64
     let groupName: String
     let groupDescription: String
     let restaurantCount: Int
